Refactored code --> LogUtil does not exist anymore. Use LogService instead
