+++ src/main/java/sese/services/BillService.java
@@ -16,7 +16,6 @@ import sese.requests.BillRequest;
 import sese.responses.BillResponse;
 import sese.responses.ReminderResponse;
 import sese.services.utils.BillCostCalculaterUtil;
-import sese.services.utils.LogUtil;
 import sese.services.utils.PdfGenerationUtil;
 import sese.services.utils.TemplateUtil;
 
@@ -37,7 +36,7 @@ public class BillService {
     private MailService mailService;
 
     @Autowired
-    private LogUtil logUtil;
+    private LogService logService;
 
     @Autowired
     public BillService(ReservationRepository reservationRepository, BillRepository billRepository, ReminderRepository reminderRepository) {
@@ -93,7 +92,7 @@ public class BillService {
         Reminder saved = reminderRepository.save(reminder);
         billRepository.save(bill);
 
-        logUtil.logAction("Eine Mahnung mit der Id '"+saved.getId()+"' wurde für die Rechnung mit der Id '"+bill.getId()+"' versendet.");
+        logService.logAction("Eine Mahnung mit der Id '"+saved.getId()+"' wurde für die Rechnung mit der Id '"+bill.getId()+"' versendet.");
 
         return new ReminderResponse(saved);
     }
@@ -151,7 +150,7 @@ public class BillService {
 
         sendBillMail(bill);
 
-        logUtil.logAction("Eine Rechnung mit der Id '"+bill.getId()+"' wurde erstellt.");
+        logService.logAction("Eine Rechnung mit der Id '"+bill.getId()+"' wurde erstellt.");
 
         return new BillResponse(bill);
     }
