+++ src/main/java/sese/services/ReservationService.java
@@ -17,7 +17,6 @@ import sese.repositories.RoomRepository;
 import sese.requests.ReservationRequest;
 import sese.requests.RoomReservationRequest;
 import sese.responses.ReservationResponse;
-import sese.services.utils.LogUtil;
 import sese.services.utils.PdfGenerationUtil;
 import sese.services.utils.TemplateUtil;
 
@@ -35,7 +34,7 @@ public class ReservationService {
     private MailService mailService;
 
     @Autowired
-    private LogUtil logUtil;
+    private LogService logService;
 
     @Autowired
     public ReservationService(CustomerRepository customerRepository, CustomerService customerService, ModelMapper modelMapper, RoomRepository roomRepository, ReservationRepository reservationRepository, MailService mailService) {
@@ -100,7 +99,7 @@ public class ReservationService {
 
         Reservation saved = reservationRepository.save(reservation);
 
-        logUtil.logAction("Eine neue Reservierung mit der Id '" + saved.getId() + "' wurde erstellt.");
+        logService.logAction("Eine neue Reservierung mit der Id '" + saved.getId() + "' wurde erstellt.");
 
         return new ReservationResponse(saved);
     }
